+++ release_assets/coverage_1.0.1.0.txt
@@ -1,0 +1,48 @@
+Name                                                 Stmts   Miss  Cover   Missing
+----------------------------------------------------------------------------------
+__init__.py                                              0      0   100%
+annflux/__init__.py                                      0      0   100%
+annflux/algorithms/__init__.py                           0      0   100%
+annflux/algorithms/embeddings.py                        20      3    85%   32-36, 47
+annflux/algorithms/feature_reconstruction_error.py      45     15    67%   40-42, 47-51, 59-69
+annflux/algorithms/most_needed.py                       51      3    94%   31, 72-73
+annflux/data/__init__.py                                 0      0   100%
+annflux/data/bombus_plant_test/__init__.py               0      0   100%
+annflux/data/bombus_plant_test/data.py                  34      2    94%   37, 48
+annflux/data/envdataset/__init__.py                      0      0   100%
+annflux/data/envdataset/data.py                         39     15    62%   35-38, 41-47, 51, 54-56, 66
+annflux/performance/__init__.py                          0      0   100%
+annflux/performance/basic.py                            68     26    62%   45-50, 56-66, 85-91, 98, 113-119, 124, 134-136
+annflux/repo_results_to_embedding.py                    64     11    83%   38-39, 47-49, 59-65, 74, 76, 96-97
+annflux/repository/__init__.py                           0      0   100%
+annflux/repository/dataset.py                           65     17    74%   27-31, 41, 43, 62, 76, 100, 104, 113, 117-118, 123, 126-128
+annflux/repository/model.py                             99     50    49%   46, 50, 54, 58, 62, 79, 89-91, 107-161, 169, 172, 179-182, 185
+annflux/repository/repository.py                       161     50    69%   37-44, 47-57, 60-64, 94, 98, 105, 151-172, 200-201, 207, 213-217, 226, 233, 287, 294, 298, 302, 317, 324
+annflux/repository/resultset.py                         94     36    62%   28-29, 91-97, 99, 108, 112, 116-117, 121, 135-151, 160-179, 182
+annflux/scripts/__init__.py                              0      0   100%
+annflux/scripts/annflux_cli.py                          59      4    93%   124-129, 138, 166
+annflux/scripts/run_tests.py                            59     20    66%   33, 40, 49, 55-63, 75, 81-83, 90-92, 97-98, 102
+annflux/shared.py                                       45      1    98%   72
+annflux/tests/__init__.py                                0      0   100%
+annflux/tests/test_cli.py                                5      0   100%
+annflux/tests/test_server.py                            43      1    98%   64
+annflux/tests/test_train.py                             27      0   100%
+annflux/tools/__init__.py                                0      0   100%
+annflux/tools/core.py                                   57      2    96%   42, 46
+annflux/tools/data.py                                  100     30    70%   158-179, 193-203, 221, 229-244
+annflux/tools/io.py                                     64     27    58%   34, 40-56, 69-74, 81, 87-95, 101
+annflux/tools/mixed.py                                  54     14    74%   30-34, 81-90
+annflux/train_indeed_image.py                          208     90    57%   40-42, 52-53, 56, 60-62, 79-86, 92, 123-147, 151-159, 171-263, 290-292
+annflux/training/__init__.py                             0      0   100%
+annflux/training/annflux/__init__.py                     0      0   100%
+annflux/training/annflux/clip.py                       242     24    90%   153, 250-254, 279-299, 309-311, 325-337
+annflux/training/annflux/clip_shared.py                 69      4    94%   105, 110-112
+annflux/training/annflux/feature_extractor.py           41      7    83%   48-50, 62, 67, 77, 82
+annflux/training/annflux/quick.py                      242     92    62%   89-90, 104-115, 194-196, 209-224, 244-249, 254-263, 276-288, 318-322, 338-346, 401, 403, 445-512
+annflux/training/tensorflow/__init__.py                  0      0   100%
+annflux/training/tensorflow/tf_backend.py               73     52    29%   38-109, 114-127, 130, 133-139
+annflux/ui/__init__.py                                   0      0   100%
+annflux/ui/basic/__init__.py                             0      0   100%
+annflux/ui/basic/run_server.py                         280    151    46%   54, 69, 124-131, 154-156, 166-168, 173-174, 177, 190-236, 241, 251-252, 259-260, 264-267, 284, 293, 302-307, 312, 317-320, 325, 330, 335, 340, 344-360, 369-423, 428-474, 498-505, 515-516, 520-526, 535
+----------------------------------------------------------------------------------
+TOTAL                                                 2408    747    69%
