--- release_assets/ruff_check_0.9.2.0.txt
@@ -1,1 +1,0 @@
-All checks passed!
