+++ release_assets/test_report_0.9.2.0.md
@@ -1,0 +1,7 @@
+| Metric | Value |
+|--------|-------|
+| Passed | 4 |
+| Failed | 0 |
+| Warnings | 11 |
+| Time | 0:03:07 |
+| Tests successful (Failed==0 and Time < 5:00) | ✅ |
