+++ release_assets/test_report_1.0.1.0.md
@@ -1,0 +1,7 @@
+| Metric | Value |
+|--------|-------|
+| Passed | 4 |
+| Failed | 0 |
+| Warnings | 11 |
+| Time | 0:03:12 |
+| Tests successful (Failed==0 and Time < 5:00) | ✅ |
