--- release_assets/tests_0.9.2.0.txt
@@ -1,15 +1,0 @@
-============================= test session starts ==============================
-platform linux -- Python 3.11.11, pytest-8.3.5, pluggy-1.5.0 -- /home/lhogeweg/miniconda3/envs/annflux2/bin/python
-cachedir: .pytest_cache
-rootdir: /home/lhogeweg/Documents/annflux_oe
-configfile: pyproject.toml
-plugins: cov-6.1.1
-collecting ... collected 2 items
-
-annflux/tests/test_cli.py::TestCli::test_cli START
-/home/lhogeweg/Documents/data/bombus-plant-test
-Testing go command
-command_="go /home/lhogeweg/annflux/data/envdataset --start_labels ['A', 'B', 'C']"
-successfully committed afe7a9ce82f5e2835df0dd66c47ec7e10e25bc294859d3592a9755cd
-Initialized AnnFlux in folder /home/lhogeweg/annflux/data/envdataset/annflux
-test_go failed
