+++ src/annflux/tools/mixed.py
@@ -13,7 +13,12 @@
 # limitations under the License.
 import logging
 import os
+import tomllib
+from importlib.metadata import version
+from pathlib import Path
 from typing import Optional, List
+
+import annflux
 
 
 def remove_sys(list_: Optional[List[str]]):
@@ -83,3 +88,18 @@ def str2bool(value, raise_exc=False):
         if raise_exc:
             raise ValueError('Expected "{}"'.format('", "'.join(_true_set | _false_set)))
     return result
+
+
+def get_version():
+    """
+    Get version of annflux
+    """
+    below_root = Path(os.path.dirname(__file__)) / ".." / ".." / ".."
+    is_package = "pyproject.toml" not in os.listdir(below_root)
+    return {
+        "version": version("annflux")
+        if is_package
+        else tomllib.load(open(below_root / "pyproject.toml", "rb"))["project"][
+            "version"
+        ]
+    }
