+++ src/annflux/ui/basic/run_server.py
@@ -17,11 +17,8 @@ import os
 import sys
 import threading
 import time
-import tomllib
 from datetime import datetime
 from functools import update_wrapper, wraps
-from importlib.metadata import version
-from pathlib import Path
 from typing import Dict, List, Optional, Tuple
 
 import flask
@@ -31,7 +28,6 @@ from flask import make_response, render_template, request, send_file
 from sklearn.ensemble import RandomForestRegressor
 from tensorflow.python.keras.callbacks import Callback
 
-import annflux
 from annflux.algorithms.embeddings import compute_tsne
 from annflux.tools.core import AnnFluxState
 from annflux.tools.data import (
@@ -39,7 +35,7 @@ from annflux.tools.data import (
     get_images_path,
     remove_uids_from_double_check,
 )
-from annflux.tools.mixed import get_logger, str2bool
+from annflux.tools.mixed import get_logger, str2bool, get_version
 from annflux.training.annflux.quick import quick_reclassification
 from annflux.training.tensorflow.tf_backend import linear_retraining
 
@@ -243,21 +239,6 @@ def label_defs_add():
 @app.route("/version")
 def version_endpoint():
     return g_version
-
-
-def get_version():
-    """
-    Get version of annflux
-    """
-    below_root = Path(os.path.dirname(annflux.__file__)) / ".." / ".."
-    is_package = "pyproject.toml" not in os.listdir(below_root)
-    return {
-        "version": version("annflux")
-        if is_package
-        else tomllib.load(open(below_root / "pyproject.toml", "rb"))["project"][
-            "version"
-        ]
-    }
 
 
 g_version = get_version()
