+++ src/annflux/ui/basic/templates/annflux_golden.html
@@ -1150,7 +1150,7 @@ limitations under the License.
 
   function forceLinearTrain() {
     $("#loading_circle").toggle();
-    statusPoll(150);
+    statusPoll(3000);
   }
 
   function highlightGalleryInMap() {
